Disable reading of compiled AST files.
I haven't thought about whether it *should* work... but at the moment,
this seems to be introducing problems when running under rbx after MRI.

+++ run.rb
@@ -131,7 +131,7 @@ convertor.virtual_root = Pathname.new( filename ).parent
 proper_filename = File.expand_path(filename).gsub('/', '\\')
 proper_filename = File.expand_path(filename) unless File.exists?( proper_filename )
 ast = nil
-if File.exists?( filename + 'c' )
+if File.exists?( filename + 'c' ) && false
 	data = nil
 	File.open( filename + 'c', 'r' ) do |file|
 		file.binmode
